ebpf: work around compiler optimization vs bpf verifier interaction
The following code causes a verifier failure on s390 with clang19:
ihl = nethdrs->iphdrs.iphdr.ihl;
if (ihl > 5) { // re-read IPv4 header if needed
    size -= bpf_core_type_size(struct iphdr);
    size += ihl * 4;
    bpf_skb_load_bytes_relative(ctx, 0, dest, size, 1);
}

ihl is a 4 bit field.  The compiler chooses to optimize the comparison
by left shifting the constant 4 bits and _then_ masking the field off
for use.  Unfortunately, the verifier saves the result and loses the
lower bounds.  Even with an explcit check for size > 0, the saved result
is still used and we get a load failure complaining about a zero length
read.

As best I can tell, the code here is structured to avoid a similar
issue.  'size' is already set to bpf_core_type_size(struct iphdr); so
this reduces to size = ihl * 4.  By shuffling the math around, we can
avoid the verifier failure.